requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
backcall==0.1.0
decorator==4.4.0
Django==2.2.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
nltk==3.4.3
numpy==1.16.4
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psycopg2-binary
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
tb-nightly==1.14.0a20190603
tensorboard==1.13.1
tensorflow==1.14.0rc1
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tflearn==0.3.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
